COMMERCE-2499 talend - component - Reorganize - rename to Operation - it refers to OpenAPI SPEC operation that equals to http method - Action comes from UI and other components we used as example - no need to replicate bad patterns :)
