+++ modules/etl/talend/talend-runtime/src/test/java/com/liferay/talend/runtime/writer/LiferayWriterTest.java
@@ -17,9 +17,9 @@ package com.liferay.talend.runtime.writer;
 import com.liferay.talend.BaseTestCase;
 import com.liferay.talend.avro.JsonObjectIndexedRecordConverter;
 import com.liferay.talend.properties.output.LiferayOutputProperties;
+import com.liferay.talend.properties.resource.Operation;
 import com.liferay.talend.runtime.LiferayRequestContentAggregatorSink;
 import com.liferay.talend.runtime.LiferaySink;
-import com.liferay.talend.tliferayoutput.Operation;
 
 import java.io.IOException;
